+++ tests/fixture/every_element/output.jsx
@@ -1,7 +1,7 @@
 const GameModeNav = ()=>{
-    return <nav data-source-path="src/every_element.jsx">
-    <ul data-source-path="src/every_element.jsx">
-      <li data-source-path="src/every_element.jsx"><Item data-source-path="src/every_element.jsx"/></li>
+    return <nav data-tsd-source="src/every_element.jsx">
+    <ul data-tsd-source="src/every_element.jsx">
+      <li data-tsd-source="src/every_element.jsx"><Item data-tsd-source="src/every_element.jsx"/></li>
     </ul>
   </nav>;
 };